home and decomp views separated, need to fix proportional df code still
