+++ src/rfd/risks/components/bond_market.py
@@ -19,7 +19,7 @@ from rfd.settings import (
     TIMES_CHOICE
 )
 
-from rfd.decomposition.pca import get_pca_components
+from rfd.tools.decomposition.pca import get_pca_components
 
 
 NAME = "Credit and Bond Market Risk"
