+++ src/rfd/risks/components/equity_market.py
@@ -15,7 +15,7 @@ from rfd.settings import (
     TIMES_CHOICE
 )
 
-from rfd.decomposition.pca import get_pca_components
+from rfd.tools.decomposition.pca import get_pca_components
 
 NAME = "Equity Market Exposure"
 COLOR = "rgb(150, 200, 150)"
