+++ src/rfd/risks/components/interest_rate.py
@@ -27,7 +27,7 @@ from rfd.settings import (
     TIMES_CHOICE
 )
 
-from rfd.decomposition.pca import get_pca_components
+from rfd.tools.decomposition.pca import get_pca_components
 
 NAME = "Yield Curve Dynamics"
 COLOR = "rgb(200, 150, 150)"
