+++ src/rfd/risks/components/market_liquidity.py
@@ -16,7 +16,7 @@ from rfd.settings import (
     TIMES_CHOICE
 )
 
-from rfd.decomposition.pca import get_pca_components
+from rfd.tools.decomposition.pca import get_pca_components
 
 NAME = "Liquidity Risk"
 COLOR = "rgb(150, 150, 200)"
