+++ src/rfd/risks/components/market_volatility.py
@@ -20,7 +20,7 @@ from rfd.settings import (
     TIMES_CHOICE
 )
 
-from rfd.decomposition.pca import get_pca_components
+from rfd.tools.decomposition.pca import get_pca_components
 
 
 NAME = "Volatility and Tail Risk"
