+++ src/rfd/server/templates/decomposition.html
@@ -1,5 +1,5 @@
 
-{% extends '/layout/top.html' %}
+{% include '/layout/top.html' %}
 {% block content %}
 
 <div class="container mt-4">
@@ -82,5 +82,5 @@
 </div>
 
 {% endblock %}
-{% extends '/layout/bottom.html' %}
+{% include '/layout/bottom.html' %}
